+++ packages/flutter/lib/rendering.dart
@@ -46,7 +46,6 @@ export 'src/rendering/layer.dart';
 export 'src/rendering/layout_helper.dart';
 export 'src/rendering/list_body.dart';
 export 'src/rendering/list_wheel_viewport.dart';
-export 'src/rendering/mouse_cursor.dart';
 export 'src/rendering/mouse_tracking.dart';
 export 'src/rendering/object.dart';
 export 'src/rendering/paragraph.dart';
@@ -65,6 +64,7 @@ export 'src/rendering/sliver_multi_box_adaptor.dart';
 export 'src/rendering/sliver_padding.dart';
 export 'src/rendering/sliver_persistent_header.dart';
 export 'src/rendering/stack.dart';
+export 'src/rendering/system_mouse_cursor.dart';
 export 'src/rendering/table.dart';
 export 'src/rendering/table_border.dart';
 export 'src/rendering/texture.dart';
