+++ packages/flutter/lib/src/material/material_state.dart
@@ -4,6 +4,7 @@
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
 
 /// Interactive states that some of the Material widgets can take on when
 /// receiving input from the user.
