+++ packages/flutter/lib/src/rendering/platform_view.dart
@@ -11,7 +11,6 @@ import 'package:flutter/services.dart';
 
 import 'box.dart';
 import 'layer.dart';
-import 'mouse_cursor.dart';
 import 'mouse_tracking.dart';
 import 'object.dart';
