+++ packages/flutter/lib/src/rendering/proxy_box.dart
@@ -14,7 +14,6 @@ import 'package:vector_math/vector_math_64.dart';
 import 'box.dart';
 import 'layer.dart';
 import 'layout_helper.dart';
-import 'mouse_cursor.dart';
 import 'mouse_tracking.dart';
 import 'object.dart';
